+++ src/test/kotlin/com/atlassian/performance/tools/infrastructure/api/distribution/PublicJiraSoftwareDistributionsIT.kt
@@ -1,7 +1,6 @@
 package com.atlassian.performance.tools.infrastructure.api.distribution
 
-import com.atlassian.performance.tools.infrastructure.toSsh
-import com.atlassian.performance.tools.sshubuntu.api.SshUbuntuContainer
+import com.atlassian.performance.tools.infrastructure.sshubuntu.SshUbuntuImage.Companion.runSoloSsh
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 
@@ -9,18 +8,15 @@ class PublicJiraSoftwareDistributionsIT {
 
     @Test
     fun shouldDownloadJiraSoftware() {
-        SshUbuntuContainer().start().use { ssh ->
-            ssh.toSsh().newConnection().use { connection ->
-                val jiraDistribution: ProductDistribution = PublicJiraSoftwareDistribution("7.2.0")
-                val targetFolder = "test"
-                connection.execute("mkdir $targetFolder")
+        runSoloSsh { ssh ->
+            val jiraDistribution: ProductDistribution = PublicJiraSoftwareDistribution("7.2.0")
+            val targetFolder = "test"
+            ssh.execute("mkdir $targetFolder")
 
-                val remoteDirectory = jiraDistribution
-                    .install(connection, targetFolder)
+            val remoteDirectory = jiraDistribution.install(ssh, targetFolder)
 
-                val directories = connection.execute("ls $remoteDirectory").output
-                assertThat(directories).contains("atlassian-jira")
-            }
+            val directories = ssh.execute("ls $remoteDirectory").output
+            assertThat(directories).contains("atlassian-jira")
         }
     }
 }