+++ src/test/kotlin/com/atlassian/performance/tools/infrastructure/api/jvm/OracleJdkIT.kt
@@ -1,19 +1,16 @@
 package com.atlassian.performance.tools.infrastructure.api.jvm
 
-import com.atlassian.performance.tools.infrastructure.toSsh
-import com.atlassian.performance.tools.sshubuntu.api.SshUbuntuContainer
+import com.atlassian.performance.tools.infrastructure.sshubuntu.SshUbuntuImage
 import org.junit.Test
 
 class OracleJdkIT {
 
     @Test
     fun shouldSupportJstatAndThreadDumps() {
-        SshUbuntuContainer().start().use { ssh ->
-            ssh.toSsh().newConnection().use { connection ->
-                val jdk = OracleJDK()
-                JstatSupport(jdk).shouldSupportJstat(connection)
-                ThreadDumpTest().shouldGatherThreadDump(jdk, connection)
-            }
+        SshUbuntuImage.runSoloSsh { ssh ->
+            val jdk = OracleJDK()
+            JstatSupport(jdk).shouldSupportJstat(ssh)
+            ThreadDumpTest().shouldGatherThreadDump(jdk, ssh)
         }
     }
 }